requirement.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-crispy-forms==1.11.1
django-filter==2.4.0
django-six==1.0.4
django-tables2==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
ez-setup==0.9
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.21.0
pandas==1.1.5
PyJWT==2.0.1
PyMuPDF==1.18.15
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tika==1.24
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
XlsxWriter==1.4.4
